rename util to utils
