+++ unbelievaboat/structures/items/StoreItemAction.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict, List, Optional
 
-from ...util.Constants import ItemActionType
+from ...utils.Constants import ItemActionType
 
 
 class StoreItemAction:
